chore: move folder to proper places
